sdpops: cast in dbg message to avoid compile warning on 32b systems
(cherry picked from commit f6921b92d382db3c0a72bd1d97eb46d1f4aee2d7)

+++ src/modules/sdpops/sdpops_mod.c
@@ -512,8 +512,9 @@ int sdp_remove_line_by_prefix(sip_msg_t* msg, str* prefix, str* media)
 
 					LM_DBG("range for media type %.*s: %ld - %ld\n",
 							sdp_stream->media.len, sdp_stream->media.s,
-							sdp_stream->raw_stream.s - body.s,
-							sdp_stream->raw_stream.s + sdp_stream->raw_stream.len - body.s
+							(long int)(sdp_stream->raw_stream.s - body.s),
+							(long int)(sdp_stream->raw_stream.s
+								+ sdp_stream->raw_stream.len - body.s)
 					);
 
 					found += sdp_remove_line_lump_by_prefix(msg,&(sdp_stream->raw_stream),prefix);
